Change paiement_due type from int to float
Updated the Paiement model to use a float for paiement_due instead of int, allowing for more precise payment amounts.

+++ app/repositories/models/paiement.py
@@ -2,7 +2,7 @@ from .base import Base
 from enums import PaiementType
 from datetime import date
 class Paiement (Base):
-    def __init__(self, id:int |None,paiement_type:int,paiement_due:int, paiement_date:date):
+    def __init__(self, id:int |None,paiement_type:int,paiement_due:float, paiement_date:date):
         super().__init__(id)
         self.paiement_type = paiement_type
         self.paiement_due = paiement_due
